Adds final, basic functionality
Started by me, fixed and completed by #chatgpt #gpt4 -- enough that I'm tagging it. I was confused about how to resolve the difference between the getOptionAttribute(for:) function setup and the need for bindings -- they don't play well together and the data layer wasn't thinking about the UI layer when it was built -- but chatGPT had a good idea for that, the use of OptionAttributeViewModel. I think the set() equivalent needs to be integrated...? So that's TBD. Next up is that potentially, some fixups for UX and UI, and strings.